moved UserTrips to models
